Migration: Fix return value of GraphBuilder.visitTypeName.
The return value of the GraphBuilder visit methods should always be
the type of the AST node assuming it's used as an expression.  When a
type name is used as an expression, its type is always a non-null
reference to the built-in type `Type`.

(Note: this wasn't causing any failures, but fixing it will allow me
to add stronger assertions in a follow-up CL.)

Change-Id: I62c72dcfc6c0f12f8d8a1a7de34a1d3f94c5e4f4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/105962
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>