[vm/kernel] Remove checks for synthetic variables in NeedsDebugStepCheck
Example: a = b ? true : false;
For a simple assignment statement with an internal (synthetic) variable on the right hand side, debugger will not allow to pause. Because DebugStepCheck will only be inserted when right hand side is a real variable.
Remove this check to enable debugger to stop on all assignments.

Bug: https://github.com/dart-lang/sdk/issues/34541
Change-Id: Iec10e5034603aa3950ad86ed5c40dc7ca1d42a03
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/107192
Commit-Queue: Zichang Guo <zichangguo@google.com>
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Reviewed-by: Ben Konyi <bkonyi@google.com>