[cfe] Store the initializer tokens in constructor builders
For const constructors, we need to compile the initializer expressions
and write them into the outline so that we can perform modular
constant evaluation.  As a first step, preserve the tokens for the
initializer expressions in const constructor builders when building
the outline.

Currently they are discarded later during outline building at the
point when we should compile them instead.

Change-Id: Ief8d94ceb752b2315982d720836496b7d597e55c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/107509
Commit-Queue: Vyacheslav Egorov <vegorov@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>