[ VM / Service ] Handle case where a JSON RPC notification results in an error
JSON RPC notifications never receive a response by design. In the VM
service, if an error needs to be returned for a request and the request
was a notification (i.e., has no ID associated with it), null is
returned. The Dart side of the service tries to create a Response object
with the null value, resulting in an exception being thrown. This change
will swallow the errors in situations where a notification RPC is
invalid.

Change-Id: If7d889150bfec51b933474f7a66fffb25e82daed
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/109200
Reviewed-by: Siva Annamalai <asiva@google.com>