[dart2js] Use Indexed{Sink/Source} for ConstantValue
This avoids naive tree-expansion of constant DAGs.

Fixes a bug in IndexedSink where multiple nodes in a tree or list can be written
as the same index, so on reading the second node is returned instead of the first.
This affects source information of inlined methods where some frames are dropped.
The .map files are about 10% larger with the frames.

Change-Id: I7e39c72510145a5ee52ea07ad2a3860bd382ee20
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/110181
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Stephen Adams <sra@google.com>