Create analysis server flag --completion-model
This flag will take a path to a folder where completion language model
files can be located. Omitting this flag will signal to analysis server
that it should not use ML ranking for code completion, and the plan is
for this capability to initially be opt-in / disabled by default.

Change-Id: Ied18cd1e378c11607778fc116a7e0c3d72091e44
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/110401
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Ari Aye <ariaye@google.com>