Bump the ABI version and oldest supported version to the same value
ABI compatibility was broken by this change:
https://dart-review.googlesource.com/c/sdk/+/108723

The change is backwards incompatible, so we can't support any versions
older than this. We should have bumped the version when the CL was
submitted, but the bot failure was misidentified as a timeout.

Bug: https://github.com/dart-lang/sdk/issues/37630
Change-Id: I3b5563ab27f40b0876b5b7682e15e2b7d7408c2f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/110541
Auto-Submit: Liam Appelbe <liama@google.com>
Reviewed-by: Siva Annamalai <asiva@google.com>
Commit-Queue: Liam Appelbe <liama@google.com>