[Test] Fixes flakiness in spawn_function_test caused by racing.
Previously the test would create and run an isolate, then checking the debugName of the isolate.
However, if the isolate ran to completion and got collected before the check happened, the debugName would be null, causing the test to fail.

The test was running at about 1.5% flakiness when run locally before.
With the change no flakiness was observed across several couple hundred runs.

Tested:
    tools/test.py -n dartk-optcounter-linux-release-x64 \
        lib_2/isolate/spawn_function_test --repeat=1000
Change-Id: I0aea6bbae08fbe321e1bea7301655ad06ec3dcd2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/108410
Reviewed-by: Ben Konyi <bkonyi@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Clement Skau <cskau@google.com>