[kernel][nnbd] Keep declared nullability on type parameter types
Programmers can use nullability modifiers on type parameters when using
them as types.  However, the actual nullability of such types is
calculated from that and also from the bound of the type parameter, and
from the promoted bound if it has it.  Previously, the final result of
the nullability calculation was stored on the type.  This CL also stores
the nullability declared by the programmer.

Change-Id: Iabbddc1e8a477678975190fa00bf702c914f7777
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/111424
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Dmitry Stefantsov <dmitryas@google.com>