[dart/vm] Rerun test on return code divergence
Rationale:
Some of the reported divergences could not be reproduced locally.
This patch adds the option to re-run tests on empty divergences
in order to triage such bugs.
Further, it adds an option to specify an existing dart program instead
of generating one randomly for testing purposes.

https://github.com/dart-lang/sdk/issues/37632
https://github.com/dart-lang/sdk/issues/37646

Change-Id: I85760dc52bfb4f3f9fa45931f94df8fa4c3c2a89
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/111240
Reviewed-by: Aart Bik <ajcbik@google.com>
Commit-Queue: Felicitas Hetzelt <felih@google.com>