[analysis_server] Fix flaky hanging in analysis server benchmark
This hanging was exposed when running analysis server benchmark on the VM
with bytecode due to different timings.

Benchmark was sending a request to analysis server and then subscribing to
a broadcast stream, which was populated from the output of analysis server
process.

Before the fix it was possible that benchmark wouldn't see
CompletionResultsParams which was posted to test.onCompletionResults
between sending a request and subscribing to test.onCompletionResults.
As a result, benchmark was waiting on a Future for a skipped
CompletionResultsParams indefinitely.

Change-Id: I6f282b504437d687ccac3e96fada201bf25dbd10
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/111484
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>