Compensate classic suggestions rather than penalizing cached ones
After a bit of manual testing it seems that subtracting 9 from
the relevanceBoost values given to the included suggestion tags
is overpenalizing them for any base suggestion set relevance
they might be holding on the client. We can compensate instead
by adding a medium size base relevance for classic completion
suggestions. The intention here is to make it so that neither
suggestion kind is universally preferred over the other for
ranking.

Change-Id: Ibaacbb35fc56c269271339df08d85b82fe6b6284
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/112150
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Ari Aye <ariaye@google.com>