[dartdevc] Fix DDK performance regression vs DDC
This explicitly types JS calls via the generic param.  Note, we also have
a type string (first arg, described in foreign_helper) that the compilers
use.  We have an analyzer hack with old DDC to use that first string arg
as a static type.  With DDK, we need to use the generic instead.  If we
don't, we end up with extraneous dynamic calls, null checks, and casts.

Exec time numbers (smaller is better):
- DDC
  - richards: 5668.555240793201 us
  - deltablue: 11736.842105263158 us
- DDK before
  - richards: 40760 us
  - deltablue: 98476.19047619049 us
- DDK after
  - richards: 1889.5184135977336 us
  - deltablue: 8432.773109243699 us

Change-Id: I720d3b40d5875d9cd6cea1c614d838f06aa86776
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/112268
Commit-Queue: Vijay Menon <vsm@google.com>
Reviewed-by: Nicholas Shahan <nshahan@google.com>