[dart2js] prepare validator for k-limiting
A few tests have huge methods where an instruction is used tens of
thousands of times. We can probably k-limit the validation in
assertion mode to a few thousand and greatly speed up those tests.

When I added k-limiting for an infeasibly large k, the program got
faster for no good reason, so I'm making this change as a baseline for
filing an issue on the VM optimizer.

Change-Id: I708d08be20652820b6ba878919ffcc6947a53b3d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/112490
Reviewed-by: Nicholas Shahan <nshahan@google.com>
Commit-Queue: Stephen Adams <sra@google.com>