[dart:io] Handle http timeout race
In http connection setup, when an explicit connection timeout is set on
the client, the onTimeout callback on the socket future caused the
future to be completed with null. If the 'then' callback on the socket
future ran before the error completer on the connection task, then
a NoSuchMethod error would result.

This CL addresses this issue by checking for null in the 'then'
callback on the socket future, and generating an appropriate
exception.

fixes https://github.com/flutter/flutter/issues/32578

Change-Id: If2e60b1355aa3d80622c644dada8a62a24469567
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/112742
Reviewed-by: Siva Annamalai <asiva@google.com>
Commit-Queue: Zach Anderson <zra@google.com>