Migration: rework handling of expressions of the form `a?.b += c`.
Previously we just had to store a single node representing the
nullability of a, since we knew we would visit `a?.b` just before
finishing up the handling of `a?.b += c`.  However, in a follow-up CL
we're going to have to change the visit order so that `a?.b` is
visited, then `c`, then `a?.b += c` can be finished up.  So we need a
map to store the association between expressions like `a?.b` and the
nullabilities of the corresponding `a`'s.

Change-Id: Ib46769e26dc63d7dc3546bddbab0d5a702a1e0a8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/113208
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>