Don't record dependencies on arguments if the instantiated class is not generic.
So, lifting a tiny bit of existing limitation of summary type inference.
Surprisingly this does not break build with summary1, but does break
one test with summary2. Probably with summary1 types are even more
broken, down to dynamic, so the types turn out to be compatible.

This passes internal tests with summary2 enabled.
https://test.corp.google.com/ui#id=OCL:263949410:BASE:263958385:1566144802009:6e0fe323

R=brianwilkerson@google.com, paulberry@google.com

Change-Id: I14a8c483778817053796fa1e047408dfb9ae3b2a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/113620
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>